qtbase: Remove the build path from the header of generated files
Source code for some examples is generated and packaged for the target.

The build path is inserted inside the comment header at the beginning
of these files. These source code files do not get installed into
the image directory and must instead be modified as part of do_package
before they are copied.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>