qtbase: Remove references to buildmachine paths
The generated qqmljsparser_p.h file contains CPP #line macros that
contain the buildpaths.  The Meta-Object-Compiler (moc) inserts
these lines to aid in mapping lines in the generated code back to
the original input lines.

It is potentially possible to suppress this by setting MOC_FLAGS
or QMAKE_MOC to include the -p flag.  However, using sed to modify
a single file is less intrusive.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>